sr(api): move find-campaign-participation-trainings in devcomp
