+++ api/src/devcomp/domain/usecases/index.js
@@ -4,6 +4,7 @@ import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
 import { repositories } from '../../infrastructure/repositories/index.js';
 import * as campaignRepository from '../../../../lib/infrastructure/repositories/campaign-repository.js';
+import * as campaignParticipationRepository from '../../../../lib/infrastructure/repositories/campaign-participation-repository.js';
 import * as knowledgeElementRepository from '../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
@@ -11,6 +12,7 @@ const path = dirname(fileURLToPath(import.meta.url));
 const dependencies = {
   ...repositories,
   campaignRepository,
+  campaignParticipationRepository,
   knowledgeElementRepository,
 };
